Adding TezosJ library to Tezos.developers website
Dear friends, I think that TezosJ, from Tezos.Rio (which I'm part of), plays an important role in Tezos ecosystem and is also a project that have been awarded with a Foundation grant and there are exchanges that use it. Please, consider letting TezosJ figure on your site. Thanks in advance, Luiz Milfont.